Update for design changes
