+++ components/rewards/RewardsHome.js
@@ -1,16 +1,8 @@
 import { FontAwesome5 } from '@expo/vector-icons';
-import { useNavigation } from '@react-navigation/native';
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  FlatList,
-  Image,
-  ScrollView,
-  TouchableOpacity,
-  View,
-} from 'react-native';
+import { FlatList, Image, ScrollView, View } from 'react-native';
 import { ProgressBar } from 'react-native-paper';
-
 import Colors from '../../constants/Colors';
 import Window from '../../constants/Layout';
 import { rewardDollarValue, rewardPointValue } from '../../constants/Rewards';
@@ -18,8 +10,8 @@ import {
   AvailableRewardsContainer,
   RewardsProgressContainer,
 } from '../../styled/rewards';
-import { ColumnContainer } from '../../styled/shared';
-import { Body, Overline, Subhead, Title } from '../BaseComponents';
+import { Body, Overline, Title } from '../BaseComponents';
+import ParticipatingStores from './ParticipatingStores';
 import RewardsCard from './RewardsCard';
 
 /**
@@ -35,31 +27,10 @@ function createList(n) {
 }
 
 function RewardsHome({ customer, participating }) {
-  const navigation = useNavigation();
   const rewardsAvailable = parseInt(customer.points, 10) / rewardPointValue;
   const pointsToNext = parseInt(customer.points, 10) % rewardPointValue;
   return (
     <ScrollView style={{ marginLeft: 16, paddingRight: 16 }}>
-      <ColumnContainer>
-        <Overline style={{ marginTop: 24, marginBottom: 4 }}>
-          Participating Stores
-        </Overline>
-        {participating.map((store) => {
-          return (
-            <TouchableOpacity key={store.id}>
-              <Subhead
-                style={{ marginLeft: 8 }}
-                onPress={() =>
-                  navigation.navigate('Stores', {
-                    currentStore: store,
-                  })
-                }>
-                {store.storeName}
-              </Subhead>
-            </TouchableOpacity>
-          );
-        })}
-      </ColumnContainer>
       <RewardsProgressContainer>
         <Overline style={{ marginTop: 24, marginBottom: 12 }}>
           Reward Progress
@@ -117,6 +88,7 @@ function RewardsHome({ customer, participating }) {
           }
         />
       </AvailableRewardsContainer>
+      <ParticipatingStores participating={participating} />
       <View style={{ maxHeight: 600, marginTop: 12 }}>
         <Image
           source={require('../../assets/images/HowItWorks.png')}
